requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.44
scikit-learn==1.5.2
xgboost==2.1.2
matplotlib==3.9.2
plotly==5.24.1
imbalanced-learn==0.12.3
seaborn==0.13.2
PyYAML==6.0.2
pytest==8.3.3